+++ ChangeLog.md
@@ -17,10 +17,19 @@ See docs/process.md for how version tagging works.
 
 Current Trunk
 -------------
+
+v1.39.8: 02/14/2020
+-------------------
 - Add LLD_REPORT_UNDEFINED option that should allow for more detailed
   diagnostics when symbols are undefined at link time.  This currently has
   some limitations and is not enabled by default. For example, EM_JS symbols
   are reported as undefined at link time, as are `__invoke_*` functions.
+- wasm2js optimizations. See binaryen#2623.
+- WebGPU Compute fixes. Simple examples now work. See #10367.
+- Many DWARF debug info fixes. Emitting of DWARF is correct as far as we know,
+  including when optimizing (a few passes are disabled for now, but almost all
+  work). We still only generate it behind the `-gforce_dwarf` flag for now,
+  though (but that should be removed soon).
 
 v1.39.7: 02/03/2020
 -------------------
@@ -33,6 +42,8 @@ v1.39.7: 02/03/2020
 - Added new linker option -s WASM=2 which produces a dual Wasm+JS build, which
   falls back to using a JavaScript version if WebAssembly is not supported in
   target browser/shell. (#10118)
+- Added new linker option -s CLOSURE_WARNINGS=quiet|warn|error that allows aborting
+  the build if Closure compiler produced any warnings.
 
 v1.39.6: 01/15/2020
 -------------------